% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-configure.R
\name{ac_init_version}
\alias{ac_init_version}
\title{Version in the \code{AC_INIT} line of configure.ac}
\usage{
ac_init_version(lines)
}
\arguments{
\item{lines}{Lines of a \code{configure.ac}.}
}
\value{
The version string of the first \verb{AC_INIT([pkg], [version])} call,
or \code{NULL} when there is none.
}
\description{
Version in the \code{AC_INIT} line of configure.ac
}
\keyword{internal}
