% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vendor-lib.R
\name{add_cargo_dependency}
\alias{add_cargo_dependency}
\title{Add a dependency to \verb{[dependencies]} in Cargo.toml}
\usage{
add_cargo_dependency(crate, version)
}
\arguments{
\item{crate}{Crate name}

\item{version}{Version spec (e.g., "*", "0.1.0")}
}
\value{
Invisibly returns TRUE if modified, FALSE if already present
}
\description{
Add a dependency to \verb{[dependencies]} in Cargo.toml
}
\keyword{internal}
