% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vendor-lib.R
\name{add_cargo_patch}
\alias{add_cargo_patch}
\title{Add a crate to \verb{[patch.crates-io]} in Cargo.toml}
\usage{
add_cargo_patch(crate, dev_path)
}
\arguments{
\item{crate}{Crate name}

\item{dev_path}{Relative path from R package root to the crate}
}
\value{
Invisibly returns TRUE if modified, FALSE if already present
}
\description{
Creates the \verb{[patch.crates-io]} section if it doesn't exist.
}
\keyword{internal}
