% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargo.R
\name{add_crate_to_workspace}
\alias{add_crate_to_workspace}
\title{Add a crate to workspace members}
\usage{
add_crate_to_workspace(workspace_toml, crate_name)
}
\arguments{
\item{workspace_toml}{Path to workspace Cargo.toml}

\item{crate_name}{Name of crate to add}
}
\value{
TRUE if successfully added, FALSE if already present
}
\description{
Modifies the workspace Cargo.toml to add a new member.
}
\keyword{internal}
