% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-native.R
\name{assert_bindgen_installed}
\alias{assert_bindgen_installed}
\title{Assert that bindgen CLI is installed}
\usage{
assert_bindgen_installed()
}
\value{
\code{NULL} invisibly when \code{bindgen} is on \code{PATH}; otherwise aborts
with the \verb{cargo install} command.
}
\description{
Assert that bindgen CLI is installed
}
\keyword{internal}
