% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bullet_created}
\alias{bullet_created}
\title{CLI bullet for file creation}
\usage{
bullet_created(path, verb = "Created")
}
\arguments{
\item{path}{Path that was created}

\item{verb}{Action verb (default "Created")}
}
\value{
\code{NULL} invisibly; called for the \code{cli} success message.
}
\description{
CLI bullet for file creation
}
\keyword{internal}
