% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargo.R
\name{cargo_clippy}
\alias{cargo_clippy}
\title{Run clippy lints}
\usage{
cargo_clippy(
  path = ".",
  release = FALSE,
  features = NULL,
  no_default_features = FALSE,
  all_features = FALSE,
  target = NULL,
  all_targets = FALSE,
  offline = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{path}{Path to the R package root, or \code{"."} to use the current directory.}

\item{release}{Logical. If TRUE, run clippy with --release.}

\item{features}{Character vector of features to activate.}

\item{no_default_features}{Logical. If TRUE, disable default features.}

\item{all_features}{Logical. If TRUE, enable all features.}

\item{target}{Character. Lint for specific target platform.}

\item{all_targets}{Logical. If TRUE, lint all targets (tests/examples/benches).}

\item{offline}{Logical. If TRUE, run without network access.}

\item{quiet}{Logical. If TRUE, suppress cargo output.}
}
\value{
Invisibly returns TRUE on success
}
\description{
Wraps \verb{cargo clippy} to run Rust lints for this package.
}
\examples{
\dontrun{
cargo_clippy()
}
}
