% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargo.R
\name{cargo_fmt}
\alias{cargo_fmt}
\title{Format Rust sources}
\usage{
cargo_fmt(path = ".", check = FALSE, all = TRUE, quiet = FALSE)
}
\arguments{
\item{path}{Path to the R package root, or \code{"."} to use the current directory.}

\item{check}{Logical. If TRUE, check formatting without modifying files.}

\item{all}{Logical. If TRUE, format all packages in the workspace.}

\item{quiet}{Logical. If TRUE, suppress cargo output.}
}
\value{
Invisibly returns TRUE on success
}
\description{
Wraps \verb{cargo fmt} to format Rust sources for this package.
}
\examples{
\dontrun{
cargo_fmt()
}
}
