% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargo.R
\name{cargo_toml_path}
\alias{cargo_toml_path}
\title{Get the Cargo.toml path for current R package}
\usage{
cargo_toml_path()
}
\value{
Path to src/rust/Cargo.toml of the active project; aborts when
the file does not exist.
}
\description{
Get the Cargo.toml path for current R package
}
\keyword{internal}
