% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_autoconf}
\alias{check_autoconf}
\title{Check autoconf availability}
\usage{
check_autoconf()
}
\value{
\code{TRUE} invisibly if autoconf is on \code{PATH}; otherwise aborts with
install hints.
}
\description{
Check autoconf availability
}
\keyword{internal}
