% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_rust}
\alias{check_rust}
\title{Check cargo/rustc availability}
\usage{
check_rust()
}
\value{
\code{TRUE} invisibly if both \code{cargo} and \code{rustc} are on \code{PATH};
otherwise aborts.
}
\description{
Check cargo/rustc availability
}
\keyword{internal}
