% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade.R
\name{configure_ac_template}
\alias{configure_ac_template}
\title{Path of the configure.ac template for a layout}
\usage{
configure_ac_template(monorepo)
}
\arguments{
\item{monorepo}{\code{TRUE} for the R package of a monorepo, \code{FALSE} for a
standalone package.}
}
\value{
Path of the installed \code{templates/monorepo/rpkg/configure.ac} or
\code{templates/rpkg/configure.ac}.
}
\description{
Path of the configure.ac template for a layout
}
\keyword{internal}
