% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_dir}
\alias{ensure_dir}
\title{Ensure directory exists}
\usage{
ensure_dir(path)
}
\arguments{
\item{path}{Path to directory}
}
\value{
Invisibly returns path
}
\description{
Ensure directory exists
}
\keyword{internal}
