% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-feature.R
\name{generate_feature_detection_rust}
\alias{generate_feature_detection_rust}
\title{Generate Rust feature detection code}
\usage{
generate_feature_detection_rust(package_name, features)
}
\arguments{
\item{package_name}{Package name}

\item{features}{Vector of feature names}
}
\value{
Character string of Rust code
}
\description{
Emits a module with one \verb{#[miniextendr]} function,
\verb{<package_rs>_enabled_features()}, that returns the names of the listed
features compiled in (checked with \verb{cfg!(feature = ...)}).
}
\keyword{internal}
