% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_package_name}
\alias{get_package_name}
\title{Get package name from current project}
\usage{
get_package_name()
}
\value{
Package name as string (the \code{Package} field of the active
project's DESCRIPTION).
}
\description{
Get package name from current project
}
\keyword{internal}
