% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_template_type}
\alias{get_template_type}
\title{Get current template type}
\usage{
get_template_type()
}
\value{
The current template type, \code{"rpkg"} (the default) or
\code{"monorepo"}, as last set by \code{\link[=set_template_type]{set_template_type()}}.
}
\description{
Get current template type
}
\keyword{internal}
