% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-lfs.R
\name{has_git_lfs}
\alias{has_git_lfs}
\title{Check if Git LFS is available}
\usage{
has_git_lfs()
}
\value{
\code{TRUE} if \verb{git lfs} is available, \code{FALSE} otherwise.
}
\description{
Check if Git LFS is available
}
\keyword{internal}
