% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-lfs.R
\name{has_git_lfs_tracking}
\alias{has_git_lfs_tracking}
\title{Check if Git LFS is tracking vendor.tar.xz}
\usage{
has_git_lfs_tracking(path = ".")
}
\arguments{
\item{path}{Path to the R package root.}
}
\value{
\code{TRUE} if \code{inst/vendor.tar.xz} is tracked by LFS, \code{FALSE} otherwise.
}
\description{
Check if Git LFS is tracking vendor.tar.xz
}
\keyword{internal}
