% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rust-source.R
\name{inline_cache_dir}
\alias{inline_cache_dir}
\title{Get inline build cache directory}
\usage{
inline_cache_dir()
}
\value{
Path to cache root: the \code{rust_source} directory under
\code{tools::R_user_dir("minirextendr", "cache")}. Not created here.
}
\description{
Get inline build cache directory
}
\keyword{internal}
