% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vendor.R
\name{miniextendr_cache_info}
\alias{miniextendr_cache_info}
\title{Show miniextendr cache info}
\usage{
miniextendr_cache_info()
}
\value{
Invisibly returns a data frame with cache info
}
\description{
Displays information about cached miniextendr archives.
}
