% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{miniextendr_pdf_document}
\alias{miniextendr_pdf_document}
\title{miniextendr-aware PDF document format}
\usage{
miniextendr_pdf_document(..., sync_args = list())
}
\arguments{
\item{...}{Arguments passed to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}}.}

\item{sync_args}{Named list of arguments passed to \code{\link[=miniextendr_sync]{miniextendr_sync()}}.}
}
\value{
An rmarkdown output format.
}
\description{
Wraps \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}} with a \code{pre_knit} hook that calls
\code{\link[=miniextendr_sync]{miniextendr_sync()}} before knitting.
}
\examples{
\dontrun{
# In YAML header:
# output: minirextendr::miniextendr_pdf_document
}
}
