% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{miniextendr_word_document}
\alias{miniextendr_word_document}
\title{miniextendr-aware Word document format}
\usage{
miniextendr_word_document(..., sync_args = list())
}
\arguments{
\item{...}{Arguments passed to \code{\link[rmarkdown:word_document]{rmarkdown::word_document()}}.}

\item{sync_args}{Named list of arguments passed to \code{\link[=miniextendr_sync]{miniextendr_sync()}}.}
}
\value{
An rmarkdown output format.
}
\description{
Wraps \code{\link[rmarkdown:word_document]{rmarkdown::word_document()}} with a \code{pre_knit} hook that calls
\code{\link[=miniextendr_sync]{miniextendr_sync()}} before knitting.
}
\examples{
\dontrun{
# In YAML header:
# output: minirextendr::miniextendr_word_document
}
}
