% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mx_desc_get_deps}
\alias{mx_desc_get_deps}
\title{Get dependencies from DESCRIPTION}
\usage{
mx_desc_get_deps(file)
}
\arguments{
\item{file}{Path to DESCRIPTION}
}
\value{
Data frame with columns: type, package, version
}
\description{
Get dependencies from DESCRIPTION
}
\keyword{internal}
