% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-detect-configure.R
\name{remove_feature_rule_from_script}
\alias{remove_feature_rule_from_script}
\title{Remove a feature rule from the script}
\usage{
remove_feature_rule_from_script(script_path, feature)
}
\arguments{
\item{script_path}{Path to detect-features.R}

\item{feature}{Feature name to remove}
}
\value{
TRUE if removed, FALSE if not found
}
\description{
Removes the single-line \code{rules[["name"]] <- function() ...} entry for the
given feature from within the \verb{## BEGIN RULES} / \verb{## END RULES} markers.
}
\keyword{internal}
