% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{render_state_path}
\alias{render_state_path}
\title{Path to render state file}
\usage{
render_state_path(path = usethis::proj_get())
}
\arguments{
\item{path}{Project root}
}
\value{
Path to \code{tools/.miniextendr/render-state.rds} under \code{path}
}
\description{
Path to render state file
}
\keyword{internal}
