% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{script_path}
\alias{script_path}
\title{Get path to bundled script}
\usage{
script_path(name)
}
\arguments{
\item{name}{Name of the script file}
}
\value{
Full path to the script
}
\description{
Get path to bundled script
}
\keyword{internal}
