% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{template_data}
\alias{template_data}
\title{Standard template data for current project}
\usage{
template_data(crate_name = NULL, package = NULL, rpkg_name = NULL)
}
\arguments{
\item{crate_name}{Optional crate name for monorepo template}

\item{package}{Optional package name override (for when DESCRIPTION doesn't exist yet)}

\item{rpkg_name}{Optional R package subdirectory name for monorepo template}
}
\value{
Named list with package, package_rs, crate_name, year, etc.
}
\description{
Standard template data for current project
}
\keyword{internal}
