% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{template_path}
\alias{template_path}
\title{Get path to package template}
\usage{
template_path(name, subdir = NULL)
}
\arguments{
\item{name}{Name of the template file (relative to template type directory)}

\item{subdir}{Optional subdirectory within the template type}
}
\value{
Full path to the template
}
\description{
For "rpkg" templates, returns templates from \verb{templates/rpkg/}.
For "monorepo" templates, returns templates from \verb{templates/monorepo/}.
}
\keyword{internal}
