% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_rust_name}
\alias{to_rust_name}
\title{Convert R package name to Rust-safe identifier}
\usage{
to_rust_name(name)
}
\arguments{
\item{name}{R package name}
}
\value{
Rust-safe name
}
\description{
Replaces dots and hyphens with underscores.
}
\keyword{internal}
