% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-rust.R
\name{use_miniextendr_document}
\alias{use_miniextendr_document}
\title{Add document.rs.in template}
\usage{
use_miniextendr_document()
}
\value{
Invisibly returns TRUE if file was created
}
\description{
Creates src/rust/document.rs.in which generates R wrapper functions
from the Rust code. This is processed by configure to create document.rs.
}
