% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-rust.R
\name{use_miniextendr_entrypoint}
\alias{use_miniextendr_entrypoint}
\title{Add entrypoint.c.in template}
\usage{
use_miniextendr_entrypoint()
}
\value{
Invisibly returns TRUE if file was created
}
\description{
Creates src/entrypoint.c.in which provides the C entry point for R
to load the shared library. Initializes panic hooks and worker thread.
}
