% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rust-source.R
\name{validate_rust_input}
\alias{validate_rust_input}
\title{Validate rust_source input}
\usage{
validate_rust_input(file, code)
}
\arguments{
\item{file}{File path or NULL}

\item{code}{Code string or NULL}
}
\value{
Character string of Rust code
}
\description{
Ensures exactly one of \code{file} or \code{code} is provided and returns the code
as a character string.
}
\keyword{internal}
