% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{write_render_state}
\alias{write_render_state}
\title{Save render state}
\usage{
write_render_state(hash, stage_run, path = usethis::proj_get())
}
\arguments{
\item{hash}{Current source hash}

\item{stage_run}{Stage that was executed}

\item{path}{Project root}
}
\value{
\code{NULL} invisibly; called for its side effect of writing the state
file (with the current time as \code{timestamp}) at \code{\link[=render_state_path]{render_state_path()}}.
}
\description{
Save render state
}
\keyword{internal}
