% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{has_miniextendr}
\alias{has_miniextendr}
\title{Check if current project has miniextendr setup}
\usage{
has_miniextendr(path = ".")
}
\arguments{
\item{path}{Path to the R package root, or \code{"."} to use the current directory.}
}
\value{
TRUE if the project appears to be a miniextendr package
}
\description{
Check if current project has miniextendr setup
}
