% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-hooks.R
\name{has_miniextendr_git_hooks}
\alias{has_miniextendr_git_hooks}
\title{Check if miniextendr git hooks are installed}
\usage{
has_miniextendr_git_hooks(path = ".")
}
\arguments{
\item{path}{Path to the R package root.}
}
\value{
Named logical vector indicating which hooks are installed.
}
\description{
Check if miniextendr git hooks are installed
}
\keyword{internal}
