% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minirextendr-package.R
\docType{package}
\name{minirextendr-package}
\alias{minirextendr}
\alias{minirextendr-package}
\title{minirextendr: Scaffold miniextendr R Packages}
\description{
Utility functions for creating and maintaining R packages that use miniextendr for Rust integration. Similar to usethis but specialized for the miniextendr framework. Provides scaffolding templates, workflow helpers, and GitHub-based vendoring of miniextendr crates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CGMossa/miniextendr}
  \item Report bugs at \url{https://github.com/CGMossa/miniextendr/issues}
}

}
\author{
\strong{Maintainer}: Mossa M. Reimert \email{mossa@sund.ku.dk}

Authors:
\itemize{
  \item Mossa M. Reimert \email{mossa@sund.ku.dk}
}

Other contributors:
\itemize{
  \item A2-AI [copyright holder, funder]
}

}
\keyword{internal}
