% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-configure.R
\name{use_miniextendr_bootstrap}
\alias{use_miniextendr_bootstrap}
\title{Add bootstrap.R to package}
\usage{
use_miniextendr_bootstrap()
}
\value{
Invisibly returns TRUE if file was created
}
\description{
Creates bootstrap.R which runs during package build when
\code{Config/build/bootstrap: TRUE} is set in DESCRIPTION.
}
