% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-configure.R
\name{use_miniextendr_config_scripts}
\alias{use_miniextendr_config_scripts}
\title{Copy config.guess and config.sub scripts}
\usage{
use_miniextendr_config_scripts()
}
\value{
Invisibly returns TRUE if files were copied
}
\description{
Copies the bundled GNU autoconf helper scripts config.guess and config.sub.
These are required for cross-compilation support.
}
