% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-configure.R
\name{use_miniextendr_makevars}
\alias{use_miniextendr_makevars}
\title{Add Makevars.in template}
\usage{
use_miniextendr_makevars()
}
\value{
Invisibly returns TRUE if file was created
}
\description{
Creates src/Makevars.in which is processed by configure to generate
the actual Makevars used during package build.
}
